refactor: SQLヘルパーをsqlHelpers.tsに抽出
gtAggregator→crossAggregatorの依存を解消するため、
共通のSQLヘルパー関数とNA_VALUEをsqlHelpers.tsに切り出し。
